[feat] 관심사에 따라 코드를 분리
