+++ js/handlers/data.js
@@ -1,0 +1,27 @@
+export const answer = {
+  "swipers[0]": [0, 1, 2, 3],
+  "swipers[1]": [0, 1],
+  "swipers[2]": [0, 1, 2, 3],
+  "swipers[3]": [0, 1],
+  "swipers[4]": [0],
+  "swipers[5]": [0, 1],
+  "swipers[6]": [0],
+  "swipers[7]": [3],
+  "swipers[8]": [0, 1, 2, 3],
+  "swipers[9]": [1, 2],
+  "swipers[10]": [1, 2],
+  "swipers[11]": [0, 1, 2, 3],
+  "swipers[12]": [1, 3],
+  "swipers[13]": [3],
+  "swipers[14]": [1],
+  "swipers[15]": [1],
+  "swipers[16]": [0, 1],
+  "swipers[17]": [0, 1, 2, 3],
+  "swipers[18]": [0, 1],
+  "swipers[19]": [2],
+  "swipers[20]": [2],
+  "swipers[21]": [0, 1, 2, 3],
+  "swipers[22]": [0, 1, 2, 3],
+  "swipers[23]": [0, 1],
+  "swipers[24]": [0, 1],
+};
