+++ js/template/data.js
@@ -90,48 +90,3 @@ export const slideData = {
     },
   ],
 };
-
-export const answer = {
-  "swipers[0]": [0, 1, 2, 3],
-  "swipers[1]": [0, 1],
-  "swipers[2]": [0, 1, 2, 3],
-  "swipers[3]": [0, 1],
-  "swipers[4]": [0],
-  "swipers[5]": [0, 1],
-  "swipers[6]": [0],
-  "swipers[7]": [3],
-  "swipers[8]": [0, 1, 2, 3],
-  "swipers[9]": [1, 2],
-  "swipers[10]": [1, 2],
-  "swipers[11]": [0, 1, 2, 3],
-  "swipers[12]": [1, 3],
-  "swipers[13]": [3],
-  "swipers[14]": [1],
-  "swipers[15]": [1],
-  "swipers[16]": [0, 1],
-  "swipers[17]": [0, 1, 2, 3],
-  "swipers[18]": [0, 1],
-  "swipers[19]": [2],
-  "swipers[20]": [2],
-  "swipers[21]": [0, 1, 2, 3],
-  "swipers[22]": [0, 1, 2, 3],
-  "swipers[23]": [0, 1],
-  "swipers[24]": [0, 1],
-};
-
-export const modalData = {
-  success: {
-    message: "축하합니다! 토끼가 당근을 발견했네요!",
-    imgSrc: "../assets/modal-success.png",
-    imgAlt: "당근을 들고 있는 토끼",
-  },
-  fail: {
-    message: "다시 한번 토끼에게 당근을 찾아주세요!",
-    imgSrc: "../assets/modal-fail.png",
-    imgAlt: "눈물 흘리는 토끼",
-  },
-};
-
-export const slideTemplateOrder = [
-  3, 1, 3, 2, 3, 2, 1, 3, 5, 4, 4, 3, 4, 3, 1, 1, 1, 5, 2, 3, 3, 3, 4, 1, 2,
-];
